QCamera, iOS: Implement new setCameraDevice behavior
The rules for how we handle camera control properties during a change
in camera-device is changing slightly. Control properties are now
transferred to the new camera device if applicable, otherwise
defaulted. This patch implements the new behavior on iOS.

Note that some existing features are not behaving as intended,
and many of them are not implemented on Android, so they
should be improved upon later.

Task-number: QTBUG-129831
Pick-to: 6.5
Change-Id: If069defc30d05a79a232e1d1e5c6379033dbb7b7
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit d9824ffd29baf4100e355034ee1f9fc83253c2a1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>