QSoundEffect: Recreate QAudioSink when setting a new QAudioDevice
We didn't recreate the existing audiosink when
QSoundEffect::setAudioDevice() was called, so audio would still be
played back by the previous audio device.

This change moves the existing audio output setup code to a new
function, updateAudioOutput(), which is then called by both
setAudioDevice() and by QSoundEffectPrivate::sampleReady(), where
it was taken from. The function replaces the old QAudioSink with
a new one based on the current audio device and loaded sample.

If previous audio device was null, updateAudioOutput() now calls
QSoundEffect::setAudioOutput(). This will update m_audioDevice
and emit audioDeviceChanged.

If QSoundEffect was already playing when setAudioOutput is called,
playback resumes from the same position when the new audio sink
has been created. This is done by converting m_offset from bytes
to a duration, and later converting and assigning it back to
m_offset based on the QAudioFormat of the newly created
QAudioBuffer.

For that reason, this change also adds an extra qMax validation
of the toWrite variable in QSoundEffectPrivate::readData, to
avoid calling memcpy with a negative third argument if an
uexpected situation or bug causes m_offset to be larger than
the QAudioBuffer size.

Lastly, this change adds an auto test that verifies/documents
the expected order of signals emitted when calling
setAuidioDevice(). To avoid having to disconnect the lambda
signals,  init(), cleanup() and initTestCase() was
modified to recreate QSoundEffect for each test function.

Fixes: QTBUG-127543
Pick-to: 6.5
Change-Id: Ibe38f20413884d7394340aab51f2f922f6e6b31e
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 83403bb62fba3658cac17faab36f909514bd26bb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>